date constraint

+++ src/main/java/bta/cabang/operasional/controller/CutiController.java
@@ -12,17 +12,18 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 
 @Controller
 public class CutiController {
     @Autowired
     private CutiService cutiService;
-
-    @Autowired
-    private UserService userService;
 
     @Autowired
     private AuthService authService;
@@ -35,7 +36,7 @@ public class CutiController {
 
         List<CutiModel> listCuti = new ArrayList<>();
         if (role == 1 || role == 2) {
-            listCuti = cutiService.getAllCuti();
+            listCuti = cutiService.getAllCutiByStatus(0);
         } else {
             listCuti = cutiService.getAllCutiByUser(id);
         }
@@ -49,23 +50,28 @@ public class CutiController {
 
     @GetMapping("/cuti/add")
     public String addCutiForm(Model model) {
+//        DateFormat df = new SimpleDateFormat("yyyy-MM-dd");
+//        Date minDate = new Date(new Date().getTime() + 86400000);
+//        String strMinDate = df.format(minDate);
+
         model.addAttribute("cuti", new CutiModel());
+//        model.addAttribute("minDate", strMinDate);
 
         return "form-addCuti";
     }
 
     @PostMapping("/cuti/add")
     public String addCutiSubmit(@ModelAttribute CutiModel cuti, RedirectAttributes redirectAttrs){
-//        try {
+        try {
             cuti.setPengaju(authService.getCurrentLoggedInUserByUsername());
             cutiService.addCuti(cuti);
             redirectAttrs.addFlashAttribute("alert", "addSuccess");
             return "redirect:/cuti";
 
-//        } catch (Exception e) {
-//            redirectAttrs.addFlashAttribute("alert", "addFail");
-//            return "redirect:/cuti";
-//        }
+        } catch (Exception e) {
+            redirectAttrs.addFlashAttribute("alert", "addFail");
+            return "redirect:/cuti";
+        }
 
     }
 
@@ -81,18 +87,25 @@ public class CutiController {
             model.addAttribute("isAbleToUpdateCuti", role == 1 || role == 2);
             return "view-cuti";
 
-        } catch (EmptyResultDataAccessException e) {
+        } catch (NoSuchElementException e) {
             redirectAttrs.addFlashAttribute("alert", "notFound");
             return "redirect:/cuti";
         }
     }
 
+    @PostMapping("/cuti/update")
+
     @GetMapping("cuti/delete/{id}")
     private String deleteCuti(@PathVariable Long id, RedirectAttributes redirectAttrs) {
         try {
-            cutiService.deleteCuti(id);
-            redirectAttrs.addFlashAttribute("alert", "delSuccess");
-            return "redirect:/cuti";
+            if (cutiService.getCutiByIdCuti(id).getStatus() == 0) {
+                cutiService.deleteCuti(id);
+                redirectAttrs.addFlashAttribute("alert", "delSuccess");
+                return "redirect:/cuti";
+            } else {
+                redirectAttrs.addFlashAttribute("alert", "delNotAllowed");
+                return "redirect:/cuti";
+            }
 
         } catch (EmptyResultDataAccessException e) {
             redirectAttrs.addFlashAttribute("alert", "notFound");
